requirements.txt
# Core dependencies
langchain
langchain-core
langchain-community
langchain-openai
langchain-experimental
openai
pydantic

# Data processing
pandas
tabulate
openpyxl  # for Excel files
numpy

# UI
streamlit

# Utils
python-dotenv
typing-extensions 